+++ src/Commands/Commands.cpp
@@ -12,6 +12,14 @@
 #include "SaveVariablesToFile.h"
 #include "ToggleSWSidebar.h"
 #include "FireTacticalSW.h"
+#include "PassiveAcquireMode.h"
+
+#include <CCINIClass.h>
+#include <InputManagerClass.h>
+#include <MouseClass.h>
+#include <WWMouseClass.h>
+
+#include <Utilities/Macro.h>
 #include <Ext/Sidebar/SWSidebar/SWSidebarClass.h>
 
 DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
@@ -22,6 +30,8 @@ DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 	MakeCommand<QuickSaveCommandClass>();
 	MakeCommand<ToggleDigitalDisplayCommandClass>();
 	MakeCommand<ToggleDesignatorRangeCommandClass>();
+	MakeCommand<AggressiveModeClass>();
+	MakeCommand<CeasefireModeClass>();
 	MakeCommand<ToggleSWSidebar>();
 
 	if (Phobos::Config::SuperWeaponSidebarCommands)