+++ src/Ext/TechnoType/Body.h
@@ -1,4 +1,4 @@
-#pragma once
+﻿#pragma once
 #include <TechnoTypeClass.h>
 
 #include <Helpers/Macro.h>
@@ -15,6 +15,13 @@
 #include <New/Type/Affiliated/DroppodTypeClass.h>
 #include <New/Type/Affiliated/TiberiumEaterTypeClass.h>
 #include <New/Type/Affiliated/CreateUnitTypeClass.h>
+
+enum class PassiveAcquireMode : BYTE
+{
+	Normal = 0,
+	Aggressive = 1,
+	Ceasefire = 2,
+};
 
 class Matrix3D;
 class ParticleSystemTypeClass;
@@ -427,6 +434,13 @@ public:
 
 		Nullable<bool> TurretResponse;
 
+		Valueable<PassiveAcquireMode> PassiveAcquireMode;
+		Valueable<bool> PassiveAcquireMode_Togglable;
+		ValueableIdx<VocClass> VoiceEnterAggressiveMode;
+		ValueableIdx<VocClass> VoiceExitAggressiveMode;
+		ValueableIdx<VocClass> VoiceEnterCeasefireMode;
+		ValueableIdx<VocClass> VoiceExitCeasefireMode;
+
 		ExtData(TechnoTypeClass* OwnerObject) : Extension<TechnoTypeClass>(OwnerObject)
 			, HealthBar_Hide { false }
 			, HealthBar_HidePips { false }
@@ -803,6 +817,14 @@ public:
 			, InfantryAutoDeploy {}
 
 			, TurretResponse {}
+				
+			, PassiveAcquireMode { PassiveAcquireMode::Normal }
+			, PassiveAcquireMode_Togglable { true }
+			, VoiceEnterAggressiveMode { -1 }
+			, VoiceExitAggressiveMode { -1 }
+			, VoiceEnterCeasefireMode { -1 }
+			, VoiceExitCeasefireMode { -1 }
+
 		{ }
 
 		virtual ~ExtData() = default;